refactor: Changed component name
