+++ src/components/RoundControlButton/RoundControlButton.tsx
@@ -1,13 +1,16 @@
 import { FaArrowLeft, FaArrowRight } from 'react-icons/fa';
 
-type PrevNextButtonTypes = 'PREV' | 'NEXT';
+type RoundControlButtonTypes = 'PREV' | 'NEXT' | 'DONE';
 
-interface PrevNextButtonProps {
-  type: PrevNextButtonTypes;
+interface RoundControlButtonProps {
+  type: RoundControlButtonTypes;
   onClick: () => void;
 }
 
-export default function PrevNextButton({ type, onClick }: PrevNextButtonProps) {
+export default function RoundControlButton({
+  type,
+  onClick,
+}: RoundControlButtonProps) {
   return (
     <button
       className="flex h-[68px] w-[200px] flex-row items-center justify-center space-x-2 rounded-full border border-neutral-300 bg-neutral-200 hover:bg-brand-main"
@@ -24,6 +27,9 @@ export default function PrevNextButton({ type, onClick }: PrevNextButtonProps) {
           <h1 className="text-[28px] font-semibold">다음 차례</h1>
           <FaArrowRight className="size-[36px]" />
         </>
+      )}
+      {type === 'DONE' && (
+        <h1 className="text-[28px] font-semibold">토론 종료</h1>
       )}
     </button>
   );
